🔥 chore: change directory structure, renaming
