+++ src/app/middlewares/verifyGithubSignature.ts
@@ -1,7 +1,7 @@
 import createError from "http-errors";
 import { createHmac } from "crypto";
 
-import Config from "../config";
+import Config from "../../config";
 import log from "@src/services/Logger";
 
 import type { RequestHandler } from "express";
