--- src/routes/repos.ts
@@ -1,14 +1,0 @@
-import { updateProjectByWebhook } from "@src/controllers/project";
-import verifyGithubSignature from "@src/middlewares/verifyGithubSignature";
-
-import { Router } from "express";
-
-const route = Router();
-
-const reposRouter = (app: Router) => {
-  app.use("/repos", route);
-
-  route.post("/hooks", verifyGithubSignature, updateProjectByWebhook);
-};
-
-export default reposRouter;
